+++ taskrs-server/src/models/request_filter.rs
@@ -6,8 +6,8 @@ pub struct RequestFilter<T> {
     pub query: Option<String>,
     pub order_by: Option<T>,
     pub order: Option<Order>,
-    pub page: Option<i32>,
-    pub limit: Option<i32>,
+    pub page: Option<i64>,
+    pub limit: Option<i64>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
